Extract common utils
